Move the last two non-queue helpers out of pg-queue
WorkLoop is a generic poll/wake/interruptible-sleep loop with a consumer
outside the queue already — realm-registry-reconciler imports it — so it goes
to work-loop.ts and stays on the barrel.

The error flattener goes to flatten-error-for-jsonb.ts and is renamed. It was
called serializableError, which is also the name of a function runtime-common
exports through a package the same callers import from, and the two do
different things: this one copies every own property so a thrown value survives
into jobs.result, while runtime-common's is card-error aware and returns
anything that is not a card error untouched. Swapping one for the other
silently changes what lands in the row, so each is now named for what it does.
It is deliberately absent from the barrel — nothing outside this package
imported it, and re-exporting would put the collision straight back.

pg-queue.ts is 788 lines, down from 1151 before this decomposition began.
